We don't care that much about locales
We only track articles on an entry+channel basis, not entry+locale+channel, since you can't have the same entry exist in the same channel for more than one locale. (No point - channels are single-language only.) So the only reason we're storing the locale in the queue is to know which locale to fetch the entry in.